Solution cleanup: interfaces into own files, add Internal namespace, etc.
- remove unused "Service References" folder
- move IContext and IVerifyContext into their own files
- make non-interface methods on Context internal instead of public
- re-order class contents to (hopefully) be easier to follow
- remove parameterless constructor for Context (not used)
- include SpecEasy.nuspec in project for easier discovery/modification
