+++ SpecEasy/DuplicateDescriptionException.cs
@@ -4,20 +4,20 @@ namespace SpecEasy
 {
     public class DuplicateDescriptionException : InvalidOperationException
     {
-        private readonly string stackTrace;
-
-        internal DuplicateDescriptionException(string typeOfDuplicate, string description, string stackTrace)
-            : base(string.Format("Failed to generate test cases; {0} description '{1}' has already been used", typeOfDuplicate, description))
-        {
-            this.stackTrace = stackTrace;
-        }
-
         public override string StackTrace
         {
             get
             {
                 return stackTrace;
             }
+        }
+
+        private readonly string stackTrace;
+
+        internal DuplicateDescriptionException(string typeOfDuplicate, string description, string stackTrace)
+            : base(string.Format("Failed to generate test cases; {0} description '{1}' has already been used", typeOfDuplicate, description))
+        {
+            this.stackTrace = stackTrace;
         }
     }
 }
