+++ SpecEasy/IContext.cs
@@ -1,0 +1,7 @@
+namespace SpecEasy
+{
+    public interface IContext
+    {
+        void Verify(Action addSpecs);
+    }
+}
