+++ SpecEasy/IVerifyContext.cs
@@ -1,0 +1,7 @@
+namespace SpecEasy
+{
+    public interface IVerifyContext
+    {
+        IVerifyContext Then(string description, Action specification);
+    }
+}
