+++ SpecEasy/Spec.TestCaseGeneration.cs
@@ -5,6 +5,7 @@ using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 using NUnit.Framework;
+using SpecEasy.Internal;
 
 namespace SpecEasy
 {
