+++ SpecEasy/Spec.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using SpecEasy.Internal;
 
 namespace SpecEasy
 {
